+++ static/templates/confirm_dialog/confirm_unsubscribe_private_stream.hbs
